services/api/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
alembic==1.11.1
python-dotenv==1.0.0
pandas==2.1.2
openpyxl==3.1.2
psycopg2-binary==2.9.7
pydantic==1.10.7
PyJWT==2.8.0
pytest==7.4.0
httpx==0.24.1
